# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dojiva API - gamified trading learning
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiMessage {
  error: string;
}

export interface LessonSummary {
  id: number;
  worldId: number;
  title: string;
  order: number;
  /** Icon key (star, chart, candle, book, trophy, chest...) */
  icon: string;
  free: boolean;
  xpReward: number;
}

export interface World {
  id: number;
  title: string;
  subtitle: string;
  emoji: string;
  order: number;
  /** True when the whole world is behind the premium paywall */
  locked: boolean;
  lessons: LessonSummary[];
}

/**
 * info = animated explanation card; quiz = multiple choice; chart-quiz = question about a rendered candlestick chart
 */
export type LessonStepType = typeof LessonStepType[keyof typeof LessonStepType];


export const LessonStepType = {
  info: 'info',
  quiz: 'quiz',
  'chart-quiz': 'chart-quiz',
} as const;

export interface LessonStep {
  id: number;
  /** info = animated explanation card; quiz = multiple choice; chart-quiz = question about a rendered candlestick chart */
  type: LessonStepType;
  prompt: string;
  /**
     * Explanation text for info steps
     * @nullable
     */
  body?: string | null;
  options?: string[];
  /** @nullable */
  correctIndex?: number | null;
  /**
     * Shown after answering
     * @nullable
     */
  explanation?: string | null;
  /**
     * JSON-encoded candle data for chart-quiz steps
     * @nullable
     */
  chart?: string | null;
}

export interface LessonDetail {
  id: number;
  worldId: number;
  title: string;
  order: number;
  icon: string;
  free: boolean;
  xpReward: number;
  steps: LessonStep[];
}

export interface LessonResult {
  score: number;
  total: number;
}

export interface Progress {
  xp: number;
  level: number;
  streak: number;
  hearts: number;
  completedLessonIds: number[];
  onboarded: boolean;
  /** @nullable */
  goal?: string | null;
  /** @nullable */
  experienceLevel?: string | null;
  markets?: string[];
  /** @nullable */
  style?: string | null;
  premium: boolean;
  /** Virtual simulator capital in EUR */
  balance: number;
}

export interface OnboardingInput {
  goal: string;
  experienceLevel: string;
  markets: string[];
  style: string;
}

export interface OkMessage {
  ok: boolean;
}

export interface Badge {
  id: number;
  code: string;
  title: string;
  description: string;
  emoji: string;
  earned: boolean;
  /** @nullable */
  earnedAt?: string | null;
}

export interface Candle {
  o: number;
  h: number;
  l: number;
  c: number;
}

export interface Scenario {
  id: number;
  market: string;
  timeframe: string;
  /** Visible candles only (the future stays hidden) */
  candles: Candle[];
  balance: number;
}

export type TradeInputDirection = typeof TradeInputDirection[keyof typeof TradeInputDirection];


export const TradeInputDirection = {
  buy: 'buy',
  sell: 'sell',
  wait: 'wait',
} as const;

export interface TradeInput {
  scenarioId: number;
  direction: TradeInputDirection;
  /** @nullable */
  entry?: number | null;
  /** @nullable */
  stopLoss?: number | null;
  /** @nullable */
  takeProfit?: number | null;
  /**
     * Percent of balance risked (default 1)
     * @nullable
     */
  riskPercent?: number | null;
  /**
     * How the user felt (confiant, stressé, neutre...)
     * @nullable
     */
  emotion?: string | null;
  /** @nullable */
  strategyId?: number | null;
}

export type TradeResultOutcome = typeof TradeResultOutcome[keyof typeof TradeResultOutcome];


export const TradeResultOutcome = {
  'take-profit': 'take-profit',
  'stop-loss': 'stop-loss',
  expired: 'expired',
  waited: 'waited',
} as const;

export interface TradeResult {
  outcome: TradeResultOutcome;
  pnl: number;
  balance: number;
  /** @nullable */
  exitPrice?: number | null;
  /** @nullable */
  riskReward: number | null;
  futureCandles: Candle[];
  /** Pedagogical feedback lines in French */
  feedback: string[];
}

export interface JournalEntry {
  id: number;
  market: string;
  direction: string;
  outcome: string;
  pnl: number;
  /** @nullable */
  riskReward?: number | null;
  /** @nullable */
  emotion?: string | null;
  /** @nullable */
  strategyName?: string | null;
  feedback?: string[];
  createdAt: string;
}

export interface JournalStats {
  totalTrades: number;
  wins: number;
  losses: number;
  winRate: number;
  /** @nullable */
  avgRiskReward: number | null;
  balance: number;
  /** @nullable */
  bestPnl: number | null;
  /** @nullable */
  worstPnl: number | null;
}

export interface Journal {
  entries: JournalEntry[];
  stats: JournalStats;
}

export interface StrategyInput {
  name: string;
  market: string;
  style: string;
  timeframe: string;
  context: string[];
  entryRules: string[];
  /** @nullable */
  stopLossRule?: string | null;
  /** @nullable */
  takeProfitRule?: string | null;
  riskPercent: number;
}

export interface Strategy {
  id: number;
  name: string;
  market: string;
  style: string;
  timeframe: string;
  context: string[];
  entryRules: string[];
  /** @nullable */
  stopLossRule?: string | null;
  /** @nullable */
  takeProfitRule?: string | null;
  riskPercent: number;
  createdAt: string;
}

export type CoachAdviceSectionsItemCategory = typeof CoachAdviceSectionsItemCategory[keyof typeof CoachAdviceSectionsItemCategory];


export const CoachAdviceSectionsItemCategory = {
  apprentissage: 'apprentissage',
  simulation: 'simulation',
  discipline: 'discipline',
} as const;

export type CoachAdviceSectionsItem = {
  category: CoachAdviceSectionsItemCategory;
  title: string;
  message: string;
};

export interface CoachAdvice {
  sections: CoachAdviceSectionsItem[];
  generatedAt: string;
}

export interface Testimonial {
  id: number;
  name: string;
  date: string;
  rating: number;
  text: string;
}