# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dojiva API - gamified trading learning
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [academy]
      summary: Get a lesson with its interactive steps
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /lessons/{id}/complete:
    post:
      operationId: completeLesson
      tags: [academy]
      summary: Mark a lesson as completed and award XP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonResult"
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /progress/onboarding:
    post:
      operationId: saveOnboarding
      tags: [academy]
      summary: Save onboarding answers and personalize the path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LessonDetail:
      type: object
      required: [id, worldId, title, order, icon, free, xpReward, steps]
      properties:
        id:
          type: integer
        worldId:
          type: integer
        title:
          type: string
        order:
          type: integer
        icon:
          type: string
        free:
          type: boolean
        xpReward:
          type: integer
        steps:
          type: array
          items:
            $ref: "#/components/schemas/LessonStep"
    LessonStep:
      type: object
      required: [id, type, prompt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [info, quiz, chart-quiz]
          description: info = animated explanation card; quiz = multiple choice; chart-quiz = question about a rendered candlestick chart
        prompt:
          type: string
        body:
          type: ["string", "null"]
          description: Explanation text for info steps
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: ["integer", "null"]
        explanation:
          type: ["string", "null"]
          description: Shown after answering
        chart:
          type: ["string", "null"]
          description: JSON-encoded candle data for chart-quiz steps
    LessonResult:
      type: object
      required: [score, total]
      properties:
        score:
          type: integer
        total:
          type: integer
    Progress:
      type: object
      required:
        [xp, level, streak, hearts, completedLessonIds, onboarded, premium, balance]
      properties:
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        hearts:
          type: integer
        completedLessonIds:
          type: array
          items:
            type: integer
        onboarded:
          type: boolean
        goal:
          type: ["string", "null"]
        experienceLevel:
          type: ["string", "null"]
        markets:
          type: array
          items:
            type: string
        style:
          type: ["string", "null"]
        premium:
          type: boolean
        balance:
          type: number
          description: Virtual simulator capital in EUR
    OnboardingInput:
      type: object
      required: [goal, experienceLevel, markets, style]
      properties:
        goal:
          type: string
        experienceLevel:
          type: string
        markets:
          type: array
          items:
            type: string
        style:
          type: string
...