# lib/api-zod/src/generated/types/progress.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dojiva API - gamified trading learning
 * OpenAPI spec version: 0.1.0
 */

export interface Progress {
  xp: number;
  level: number;
  streak: number;
  hearts: number;
  completedLessonIds: number[];
  onboarded: boolean;
  /** @nullable */
  goal?: string | null;
  /** @nullable */
  experienceLevel?: string | null;
  markets?: string[];
  /** @nullable */
  style?: string | null;
  premium: boolean;
  /** Virtual simulator capital in EUR */
  balance: number;
}